refactor: #2097 demo / 本番 child home UI を真に 1 ファイル描画に統合 (shim 撤廃、6 回目指摘の構造的解決)
過去 5 回 (#531 / #561 / #562 / #563 / #566 / #2069) 同種要求があり、各回 Dev/Reviewer Agent は
「Tier N で統合」「scope 外」「UI 等価性 SS 維持」と判断して shim 実装 (Service DI 足場 + 派生コンポーネント)
で完了報告してきた。実体として demo / 本番の並行実装 2 系統は構造的に残存していた。
本 PR でその shim を撤廃し、研究資料 §3「ルーティングレベルでの分離」準拠の真の SSOT 化を完遂する。

## 実装

- `DashboardView.svelte` を本番 + demo 両方の SSOT に昇格 (1 ファイル描画)
  - 5 年齢モード分岐 (baby/preschool/elementary/junior/senior)
  - pin / mission badge / xp animation / baby inline form / event badge / sibling ranking
  - dialog FSM / OverlaysSection / Adventure / ParentMessage / SiblingCelebration / Cheer
  - login stamp 自動 claim / Birthday banner / Tutorial hint / Error toast
- `ProdDashboardSections.svelte` (322 行) を `git rm` で削除
- 本番 `+page.svelte` を 986 → 49 行の薄ラッパに刷新
- demo `+page.svelte` を 51 → 33 行の薄ラッパに刷新 (本番と全く同じ DashboardView を呼ぶ)
- demo `+layout.server.ts` + `home/+page.server.ts` を拡張し本番 layout / load 戻り値と同等のフィールド
  (planLimits / isPremium / categoryXp / activeEvents / activeChallenges / siblingRanking 等) を mock 提供
- ChildDashboardService interface (ADR-0046) はそのまま流用

## 機械検証 AC

- AC1: src/lib/features/child-home/components/ 配下のダッシュボード描画ファイル数 = 1 (DashboardView.svelte のみ)
- AC2: 本番 / demo 両 +page.svelte が同じ DashboardView を参照 (grep 結果 = 2)
- AC3: ProdDashboardSections.svelte 不在を file system で確認

## 検証

- biome check: PASS
- svelte-check: 0 ERRORS (warnings は別ファイル admin/activities/edit、本 PR とは無関係)
- vitest run: 4573/4573 PASS (full suite、退行 0)
- 新規テスト: sessionStorage 隔離 (研究資料 §4-1) + interface 整合性 6 件追加 → 39/39 PASS

## 設計書更新

- ADR-0046 結果欄に #2097 の真の共通化完了を追記
- docs/design/parallel-implementations.md §3 から child home エントリを撤去 (統合完了表記でなく削除)
- docs/design/06-UI設計書.md §4.1 に SSOT 実装ファイルパスを明記

## 逆方向統合の禁忌

本番固有 feature (pin / xp animation 等) は demo に逆輸入しない誘惑を排除した。demo に表示されるのが期待動作。
LP SS が本番 SS に寄って見えるのが正常 (Issue #2097 §統合方針)。

PO 動作確認手順は PR body に記載。Draft 据置で PO レビュー待ち。

Closes #2097
