+++ docs/design/06-UI設計書.md
@@ -139,6 +139,19 @@
 
 ### 4.1 子供用ホーム画面 (S-01)
 
+> **SSOT 実装ファイル** (Issue #2097 / ADR-0046、2026-05-14 真の共通化完了):
+> `src/lib/features/child-home/components/DashboardView.svelte` の **1 ファイル**が
+> **本番 (`/(child)/[uiMode]/home`) と demo (`/demo/(child)/[mode]/home`) の両方を描画**する。
+> 本番 `+page.svelte` も demo `+page.svelte` も薄いラッパで、`<DashboardView {data} />` のみ呼ぶ。
+>
+> 5 年齢モード分岐 / pin / mission badge / xp animation / baby inline form / event badge /
+> sibling ranking / dialog FSM / OverlaysSection 等の全 feature を 1 ファイルに集約済。
+> 仕様変更時はこのファイル 1 つを編集すれば本番 / demo 両方に同時反映される (並行実装 0 件)。
+>
+> Service Interface (`ChildDashboardService`) + Svelte 5 Context DI で本番 (`ProductionDashboardService` /
+> REST `/api/v1/...` fetch) と demo (`DemoDashboardService` / sessionStorage + `$state`) を切り替える。
+
+
 ```
 ┌─────────────────────────────────┐
 │  [👤 名前]         ⭐ 1,250P    │  ← ヘッダー（テーマカラー）
