+++ src/routes/demo/(child)/+layout.server.ts
@@ -3,6 +3,21 @@ import { DEMO_CHILDREN } from '$lib/server/demo/demo-data.js';
 import { getDemoChildLayoutData } from '$lib/server/demo/demo-service.js';
 import type { LayoutServerLoad } from './$types';
 
+/**
+ * Issue #2097: 真の child home UI 統合 (6 回目指摘)
+ *
+ * demo 側 layout も本番側 (`(child)/+layout.server.ts`) と同等のフィールドを返す。
+ * 本番 layout が返す `planLimits` / `isPremium` / `milestones` / `stampProgress` /
+ * `stampCard` / `planTier` を demo 側も提供することで、共通 `DashboardView.svelte`
+ * が 1 つの prop 契約で本番 / demo 両方を描画できる。
+ *
+ * demo は実装の事実上 free プラン体験を提供する想定。`planLimits` は free 相当の
+ * 既定値、`isPremium` は false、`milestones` は空配列 (screenshot=all モード時のみ
+ * +layout.svelte で固定 milestone を別途生成)。
+ */
+
+import { getPlanLimits, isPaidTier } from '$lib/server/services/plan-limit-service';
+
 export const load: LayoutServerLoad = async ({ url }) => {
 	// childId from query string (set when selecting a child from demo landing)
 	const childIdParam = url.searchParams.get('childId');
@@ -22,9 +37,21 @@ export const load: LayoutServerLoad = async ({ url }) => {
 	}
 
 	// Fallback to switch page if no child found
-	if (!childId) {
-		return getDemoChildLayoutData(0);
-	}
+	const layoutData = childId ? getDemoChildLayoutData(childId) : getDemoChildLayoutData(0);
 
-	return getDemoChildLayoutData(childId);
+	// #2097: 本番 layout が返す追加フィールドを demo でも提供
+	// demo は free 相当の体験。
+	const planTier = 'free' as const;
+	const planLimits = getPlanLimits(planTier);
+	const isPremium = isPaidTier(planTier);
+
+	return {
+		...layoutData,
+		planTier,
+		planLimits,
+		isPremium,
+		stampProgress: null,
+		stampCard: null,
+		milestones: [],
+	};
 };
